# backend/openapi.yaml
openapi: 3.1.0
info:
  title: DIP-easy
  version: 0.1.0
paths:
...
  /api/v1/products/:
    get:
      tags:
      - products
      summary: Read Products
      description: Retrieve products.
      operationId: read_products_api_v1_products__get
      parameters:
      - required: false
        schema:
          type: integer
          title: Skip
          default: 0
        name: skip
        in: query
      - required: false
        schema:
          type: integer
          title: Limit
          default: 100
        name: limit
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Product'
                type: array
                title: Response Read Products Api V1 Products  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - products
      summary: Create Product
      description: Create new product.
      operationId: create_product_api_v1_products__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/products/{product_id}:
    get:
      tags:
      - products
      summary: Read Product
      description: Get product by ID.
      operationId: read_product_api_v1_products__product_id__get
      parameters:
      - required: true
        schema:
          type: string
          format: uuid
          title: Product Id
        name: product_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - products
      summary: Update Product
      description: Update a product.
      operationId: update_product_api_v1_products__product_id__put
      parameters:
      - required: true
        schema:
          type: string
          format: uuid
          title: Product Id
        name: product_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    delete:
      tags:
      - products
      summary: Delete Product
      description: Delete a product.
      operationId: delete_product_api_v1_products__product_id__delete
      parameters:
      - required: true
        schema:
          type: string
          format: uuid
          title: Product Id
        name: product_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/products/{product_id}/submit:
    put:
      tags:
      - products
      summary: Submit Product
      description: Validate and mark product as VALIDATED
      operationId: submit_product_api_v1_products__product_id__submit_put
      parameters:
      - required: true
        schema:
          type: string
          format: uuid
          title: Product Id
        name: product_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    CompatibilityTest:
      properties:
        test_type:
          type: string
          title: Test Type
        test_date:
          type: string
          format: date-time
          title: Test Date
        result:
          type: string
          title: Result
        notes:
          type: string
          title: Notes
        id:
          type: string
          format: uuid
          title: Id
        product_id:
          type: string
          format: uuid
          title: Product Id
      type: object
      required:
      - test_type
      - test_date
      - result
      - id
      - product_id
      title: CompatibilityTest
    CompatibilityTestCreate:
      properties:
        test_type:
          type: string
          title: Test Type
        test_date:
          type: string
          format: date-time
          title: Test Date
        result:
          type: string
          title: Result
        notes:
          type: string
          title: Notes
      type: object
      required:
      - test_type
      - test_date
      - result
      title: CompatibilityTestCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Ingredient:
      properties:
        ingr_nom_inci:
          type: string
          title: Ingr Nom Inci
        ingr_fonction:
          type: string
          title: Ingr Fonction
        ingr_pourcent_min:
          type: number
          title: Ingr Pourcent Min
        ingr_pourcent_max:
          type: number
          title: Ingr Pourcent Max
        ingr_cas:
          type: string
          title: Ingr Cas
        ingr_provenance:
          type: string
          title: Ingr Provenance
        ingr_specif:
          type: string
          title: Ingr Specif
        id:
          type: string
          format: uuid
          title: Id
        product_id:
          type: string
          format: uuid
          title: Product Id
      type: object
      required:
      - ingr_nom_inci
      - ingr_fonction
      - id
      - product_id
      title: Ingredient
    IngredientCreate:
      properties:
        ingr_nom_inci:
          type: string
          title: Ingr Nom Inci
        ingr_fonction:
          type: string
          title: Ingr Fonction
        ingr_pourcent_min:
          type: number
          title: Ingr Pourcent Min
        ingr_pourcent_max:
          type: number
          title: Ingr Pourcent Max
        ingr_cas:
          type: string
          title: Ingr Cas
        ingr_provenance:
          type: string
          title: Ingr Provenance
        ingr_specif:
          type: string
          title: Ingr Specif
      type: object
      required:
      - ingr_nom_inci
      - ingr_fonction
      title: IngredientCreate
...
    Product:
      properties:
        nom_commercial:
          type: string
          title: Nom Commercial
        fournisseur:
          type: string
          title: Fournisseur
        ref_formule:
          type: string
          title: Ref Formule
        ref_produit:
          type: string
          title: Ref Produit
        date_mise_marche:
          type: string
          format: date
          title: Date Mise Marche
        resp_mise_marche:
          type: string
          title: Resp Mise Marche
        faconnerie:
          type: string
          title: Faconnerie
        pc_ph:
          type: number
          title: Pc Ph
        pc_densite:
          type: number
          title: Pc Densite
        pc_organoleptiques:
          type: string
          title: Pc Organoleptiques
        pao_theorique:
          type: string
          title: Pao Theorique
        pao_valide:
          type: string
          title: Pao Valide
        util_descr:
          type: string
          title: Util Descr
        util_precaution:
          type: string
          title: Util Precaution
        tox_nanomaterials:
          type: boolean
          title: Tox Nanomaterials
          default: false
        ei_signalements:
          type: string
          title: Ei Signalements
        autres_tests:
          type: string
          title: Autres Tests
        effets_revendiques:
          type: string
          title: Effets Revendiques
        status:
          $ref: '#/components/schemas/ProductStatus'
        ingredients:
          items:
            $ref: '#/components/schemas/Ingredient'
          type: array
          title: Ingredients
          default: []
        stability_tests:
          items:
            $ref: '#/components/schemas/StabilityTest'
          type: array
          title: Stability Tests
          default: []
        compatibility_tests:
          items:
            $ref: '#/components/schemas/CompatibilityTest'
          type: array
          title: Compatibility Tests
          default: []
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - status
      - id
      - created_at
      - updated_at
      title: Product
    ProductCreate:
      properties:
        nom_commercial:
          type: string
          title: Nom Commercial
        fournisseur:
          type: string
          title: Fournisseur
        ref_formule:
          type: string
          title: Ref Formule
        ref_produit:
          type: string
          title: Ref Produit
        date_mise_marche:
          type: string
          format: date
          title: Date Mise Marche
        resp_mise_marche:
          type: string
          title: Resp Mise Marche
        faconnerie:
          type: string
          title: Faconnerie
        pc_ph:
          type: number
          title: Pc Ph
        pc_densite:
          type: number
          title: Pc Densite
        pc_organoleptiques:
          type: string
          title: Pc Organoleptiques
        pao_theorique:
          type: string
          title: Pao Theorique
        pao_valide:
          type: string
          title: Pao Valide
        util_descr:
          type: string
          title: Util Descr
        util_precaution:
          type: string
          title: Util Precaution
        tox_nanomaterials:
          type: boolean
          title: Tox Nanomaterials
          default: false
        ei_signalements:
          type: string
          title: Ei Signalements
        autres_tests:
          type: string
          title: Autres Tests
        effets_revendiques:
          type: string
          title: Effets Revendiques
        status:
          allOf:
          - $ref: '#/components/schemas/ProductStatus'
          default: DRAFT
        ingredients:
          items:
            $ref: '#/components/schemas/IngredientCreate'
          type: array
          title: Ingredients
          default: []
        stability_tests:
          items:
            $ref: '#/components/schemas/StabilityTestCreate'
          type: array
          title: Stability Tests
          default: []
        compatibility_tests:
          items:
            $ref: '#/components/schemas/CompatibilityTestCreate'
          type: array
          title: Compatibility Tests
          default: []
      type: object
      title: ProductCreate
    ProductStatus:
      type: string
      enum:
      - DRAFT
      - VALIDATED
      title: ProductStatus
      description: An enumeration.
    ProductUpdate:
      properties:
        nom_commercial:
          type: string
          title: Nom Commercial
        fournisseur:
          type: string
          title: Fournisseur
        ref_formule:
          type: string
          title: Ref Formule
        ref_produit:
          type: string
          title: Ref Produit
        date_mise_marche:
          type: string
          format: date
          title: Date Mise Marche
        resp_mise_marche:
          type: string
          title: Resp Mise Marche
        faconnerie:
          type: string
          title: Faconnerie
        pc_ph:
          type: number
          title: Pc Ph
        pc_densite:
          type: number
          title: Pc Densite
        pc_organoleptiques:
          type: string
          title: Pc Organoleptiques
        pao_theorique:
          type: string
          title: Pao Theorique
        pao_valide:
          type: string
          title: Pao Valide
        util_descr:
          type: string
          title: Util Descr
        util_precaution:
          type: string
          title: Util Precaution
        tox_nanomaterials:
          type: boolean
          title: Tox Nanomaterials
          default: false
        ei_signalements:
          type: string
          title: Ei Signalements
        autres_tests:
          type: string
          title: Autres Tests
        effets_revendiques:
          type: string
          title: Effets Revendiques
        status:
          allOf:
          - $ref: '#/components/schemas/ProductStatus'
          default: DRAFT
        ingredients:
          items:
            $ref: '#/components/schemas/IngredientCreate'
          type: array
          title: Ingredients
          default: []
        stability_tests:
          items:
            $ref: '#/components/schemas/StabilityTestCreate'
          type: array
          title: Stability Tests
          default: []
        compatibility_tests:
          items:
            $ref: '#/components/schemas/CompatibilityTestCreate'
          type: array
          title: Compatibility Tests
          default: []
      type: object
      title: ProductUpdate
    StabilityTest:
      properties:
        test_type:
          type: string
          title: Test Type
        test_date:
          type: string
          format: date-time
          title: Test Date
        result:
          type: string
          title: Result
        notes:
          type: string
          title: Notes
        id:
          type: string
          format: uuid
          title: Id
        product_id:
          type: string
          format: uuid
          title: Product Id
      type: object
      required:
      - test_type
      - test_date
      - result
      - id
      - product_id
      title: StabilityTest
    StabilityTestCreate:
      properties:
        test_type:
          type: string
          title: Test Type
        test_date:
          type: string
          format: date-time
          title: Test Date
        result:
          type: string
          title: Result
        notes:
          type: string
          title: Notes
      type: object
      required:
      - test_type
      - test_date
      - result
      title: StabilityTestCreate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/token